fix(presence,scheduling,entities): repair test regressions from #527
The homogeneity audit (#527) broke 14 tests across three packages:

- react-scheduling / react-entities: query-key builders were extracted
  into hooks/query-keys.ts but the hook files no longer re-exported them,
  so tests importing them from the hook path failed with
  'is not a function'. Re-export the builders from the hook files, matching
  the fix already applied to react-authorization / react-authentication-api-keys.

- presence: chunking helper was moved into @granit/presence as a
  fetch*-prefixed function (fetchBatchPresence) that calls getBatchPresence
  intra-module, so the hook test could no longer mock the network seam, and
  the name violates the get*/list* verb convention. Move chunking back into
  the useBatchPresence hook (getBatchPresenceChunked) so each chunk hits the
  mockable getBatchPresence primitive; drop fetchBatchPresence from core.
